Savagely split the dns/registrar stuff in a new dns.py
